+++ ha-dashboard/src/main/java/jp/co/ha/dashboard/calorie/controller/CalorieCalcController.java
@@ -16,7 +16,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 
 import jp.co.ha.business.component.CalorieApiComponent;
 import jp.co.ha.business.component.UserComponent;
-import jp.co.ha.business.db.crud.read.UserSearchService;
 import jp.co.ha.business.dto.CalorieCalcDto;
 import jp.co.ha.business.exception.BusinessException;
 import jp.co.ha.business.healthInfo.type.GenderType;
@@ -43,9 +42,6 @@ public class CalorieCalcController implements BaseWebController {
     /** カロリー計算APIComponent */
     @Autowired
     private CalorieApiComponent calorieCalcComponent;
-    /** ユーザ情報検索サービス */
-    @Autowired
-    private UserSearchService userSearchService;
     /** ユーザ関連Component */
     @Autowired
     private UserComponent userComponent;
@@ -67,7 +63,7 @@ public class CalorieCalcController implements BaseWebController {
 
         Long seqUserId = sessionComponent
                 .getValue(request.getSession(), "seqUserId", Long.class).get();
-        Optional<User> user = userSearchService.findById(seqUserId);
+        Optional<User> user = userComponent.findById(seqUserId);
 
         form.setAge(userComponent.getAge(user.get().getBirthDate()));
         form.setGender(user.get().getGenderType().toString());
@@ -117,7 +113,7 @@ public class CalorieCalcController implements BaseWebController {
         Long seqUserId = sessionComponent
                 .getValue(request.getSession(), "seqUserId", Long.class).get();
 
-        Optional<User> user = userSearchService.findById(seqUserId);
+        Optional<User> user = userComponent.findById(seqUserId);
 
         dto.setGenderType(GenderType.of(user.get().getGenderType().intValue()));
