+++ ha-dashboard/src/main/java/jp/co/ha/dashboard/healthinfo/service/impl/HealthInfoMailServiceImpl.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Service;
 import jp.co.ha.business.api.aws.AwsSesComponent;
 import jp.co.ha.business.api.aws.AwsSesComponent.MailTemplateKey;
 import jp.co.ha.business.api.healthinfoapp.response.HealthInfoRegistApiResponse;
-import jp.co.ha.business.db.crud.read.UserSearchService;
+import jp.co.ha.business.component.UserComponent;
 import jp.co.ha.business.io.file.properties.HealthInfoProperties;
 import jp.co.ha.common.exception.BaseException;
 import jp.co.ha.common.util.DateTimeUtil;
@@ -24,9 +24,9 @@ import jp.co.ha.dashboard.healthinfo.service.HealthInfoMailService;
 @Service
 public class HealthInfoMailServiceImpl implements HealthInfoMailService {
 
-    /** ユーザ情報検索サービス */
+    /** ユーザComponent */
     @Autowired
-    private UserSearchService userSearchService;
+    private UserComponent userComponent;
     /** AWS-SES Component */
     @Autowired
     private AwsSesComponent ses;
@@ -39,7 +39,7 @@ public class HealthInfoMailServiceImpl implements HealthInfoMailService {
             throws BaseException {
 
         Long seqUserId = apiResponse.getAccount().getSeqUserId();
-        String to = userSearchService.findById(seqUserId).get().getMailAddress();
+        String to = userComponent.findById(seqUserId).get().getMailAddress();
         String titleText = "健康情報登録完了メール" + DateTimeUtil.toString(
                 DateTimeUtil.getSysDate(), DateTimeUtil.DateFormatType.YYYYMMDD_NOSEP);
         Map<String, String> bodyMap = new HashMap<>();
